requirements.txt
requests
pandas
clickhouse-connect==0.6.8
duckdb==0.10.2
psycopg2-binary
fastapi
uvicorn
pydantic
pydantic-settings
streamlit
plotly
numpy